Add graph view to sd card window
